Moved native functions to own file. Fix object type equality. Fix noop overwrite on calls to empty functions

+++ src/lexer.rs
@@ -5,7 +5,7 @@ use crate::token::{Token, TokenKind, Span};
 #[derive(Debug)]
 pub struct Lexer {
 	pub source: String,
-	line: u64,
+	line: u32,
 	column: u32,
 	pos: usize,
 }
